Rename commands=>cmd
