+++ exp/services/recoverysigner/cmd/serve.go
@@ -1,4 +1,4 @@
-package commands
+package cmd
 
 import (
 	"go/types"
