+++ exp/services/recoverysigner/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
-	"github.com/stellar/go/exp/services/recoverysigner/internal/commands"
+	"github.com/stellar/go/exp/services/recoverysigner/cmd"
 	supportlog "github.com/stellar/go/support/log"
 )
 
@@ -19,7 +19,7 @@ func main() {
 		},
 	}
 
-	rootCmd.AddCommand((&commands.ServeCommand{Logger: logger}).Command())
+	rootCmd.AddCommand((&cmd.ServeCommand{Logger: logger}).Command())
 
 	err := rootCmd.Execute()
 	if err != nil {
